% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToAminoAcid.R
\name{ConvertToAminoAcid}
\alias{ConvertToAminoAcid}
\title{Convert a given DNA sequence or a RNA sequence to an animo acid sequence}
\usage{
ConvertToAminoAcid(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
The input DNA or RNA sequence will be converted to an amino acid sequence
}
\description{
Convert a given DNA sequence or a RNA sequence to an animo acid sequence
}
\examples{
x <- "AATGTTAT"
ConvertToAminoAcid(x)
}
